+++ examples/_shared/Flavors.php
@@ -11,8 +11,10 @@ declare(strict_types=1);
 
 namespace Examples\Shared;
 
+use Examples\ApiResources\ExampleServiceProvider as ApiResources;
 use Examples\Combined\ExampleServiceProvider as Combined;
 use Examples\FormRequests\ExampleServiceProvider as FormRequests;
+use Examples\Fractal\ExampleServiceProvider as Fractal;
 use Examples\QueryBuilder\ExampleServiceProvider as QueryBuilder;
 use Examples\SpatieData\ExampleServiceProvider as SpatieData;
 use Examples\Vanilla\ExampleServiceProvider as Vanilla;
@@ -34,6 +36,8 @@ final class Flavors
             'vanilla'       => Vanilla::class,
             'form-requests' => FormRequests::class,
             'spatie-data'   => SpatieData::class,
+            'api-resources' => ApiResources::class,
+            'fractal'       => Fractal::class,
             'query-builder' => QueryBuilder::class,
             'combined'      => Combined::class,
         ];